Refactoring of the translation strings in atomic components
