+++ resources/js/components/atoms/buttons/ButtonLoadMore.vue
@@ -3,7 +3,7 @@
     class="text-xl mt-10 h-20"
     v-on="$listeners"
   >
-    {{ loadMore }}
+    {{ __('generic.load_more') }}
   </ButtonArrow>
 </template>
 
@@ -13,12 +13,6 @@ import ButtonArrow from './ButtonArrow'
 export default {
   components: {
     ButtonArrow
-  },
-  props: {
-    loadMore: {
-      type: String,
-      default: 'Load more'
-    }
   }
 }
 </script>
