+++ resources/lang/en/documents.php
@@ -5,8 +5,6 @@ return [
     'created' => 'Created',
     'preview' => 'Preview',
     'download' => 'Download',
-    'cancel_filters' => 'Cancel filters',
-    'load_more' => 'Load more',
     'no_documents' => 'No documents found',
     'document_type' => 'Document type',
     'topic' => 'Topic',
