+++ resources/lang/en/generic.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    'cancel_filters' => 'Cancel filters',
+    'load_more' => 'Load more',
+];
