+++ resources/lang/en/post.php
@@ -9,7 +9,5 @@ return [
     'navigate_back' => 'Back to all information',
     'related' => 'Related',
     'not_published' => 'Draft',
-    'cancel_filters' => 'Cancel filters',
     'no_posts' => 'No posts found',
-    'load_more' => 'Load more',
 ];
