+++ resources/lang/sk/documents.php
@@ -5,8 +5,6 @@ return [
     'created' => 'Vytvorený',
     'preview' => 'Náhľad',
     'download' => 'Stiahnuť',
-    'cancel_filters' => 'Zrušiť filter',
-    'load_more' => 'Načítať ďalšie',
     'no_documents' => 'Nenašli sa žiadne dokumenty',
     'document_type' => 'Typ dokumentu',
     'topic' => 'Téma',
