+++ resources/lang/sk/generic.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    'cancel_filters' => 'Zrušiť filter',
+    'load_more' => 'Načítať ďalšie',
+];
