+++ resources/lang/sk/post.php
@@ -9,7 +9,5 @@ return [
     'navigate_back' => 'Späť na všetky informácie',
     'related' => 'Súvisiace',
     'not_published' => 'Návrh',
-    'cancel_filters' => 'Zrušiť filter',
     'no_posts' => 'Nenašli sa žiadne články',
-    'load_more' => 'Načítať ďalšie',
 ];
